update room connection logic for Y-shaped map
This modifies several statements describing the connecting logic of various rooms. It radically changes the visual appearance of the map, due to multiple changes all at once.

A better demonstration of the debugging process might be to start by adding the Plaza inside Citadel City first, and then walk through each commit reconnecting each Gate to the Plaza. Fine tuning of the layout will continue over many iterations. But we are Wizards, and we can move entire mountains with mere keystrokes.